ssz_encoding/types/ssz.go
package types

import (
	"fmt"
	ssz "github.com/ferranbt/fastssz"
)

// SSZ32Bytes --
type SSZ32Bytes [32]byte

func (b SSZ32Bytes) HashTreeRoot() ([32]byte, error) {
	return ssz.HashWithDefaultHasher(b)
}

func (b SSZ32Bytes) GetTree() (*ssz.Node, error) {
	return ssz.ProofTree(b)
}

func (b SSZ32Bytes) HashTreeRootWith(hh ssz.HashWalker) error {
	indx := hh.Index()
	hh.PutBytes(b[:])
	hh.Merkleize(indx)
	return nil
}

// UnmarshalSSZ --
func (b SSZ32Bytes) UnmarshalSSZ(buf []byte) error {
	if len(buf) != b.SizeSSZ() {
		return fmt.Errorf("expected buffer of length %d receiced %d", b.SizeSSZ(), len(buf))
	}
	copy(b[:], buf[:])
	return nil
}

// MarshalSSZTo --
func (b SSZ32Bytes) MarshalSSZTo(dst []byte) ([]byte, error) {
	return append(dst, b[:]...), nil
}

// MarshalSSZ --
func (b SSZ32Bytes) MarshalSSZ() ([]byte, error) {
	return b[:], nil
}

// SizeSSZ returns the size of the serialized object.
func (b SSZ32Bytes) SizeSSZ() int {
	return 32
}

// Contributions --
type Contributions []*Contribution

func (c *Contributions) HashTreeRoot() ([32]byte, error) {
	return ssz.HashWithDefaultHasher(c)
}

func (c *Contributions) GetTree() (*ssz.Node, error) {
	return ssz.ProofTree(c)
}

func (c *Contributions) HashTreeRootWith(hh ssz.HashWalker) error {
	// taken from https://github.com/prysmaticlabs/prysm/blob/develop/encoding/ssz/htrutils.go#L97-L119
	subIndx := hh.Index()
	num := uint64(len(*c))
	if num > 13 {
		return ssz.ErrIncorrectListSize
	}
	for _, elem := range *c {
		{
			if err := elem.HashTreeRootWith(hh); err != nil {
				return err
			}
		}
	}
	hh.MerkleizeWithMixin(subIndx, num, 13)
	return nil
}

// UnmarshalSSZ --
func (c *Contributions) UnmarshalSSZ(buf []byte) error {
	num, err := ssz.DecodeDynamicLength(buf, 13)
	if err != nil {
		return err
	}
	*c = make(Contributions, num)

	return ssz.UnmarshalDynamic(buf, num, func(indx int, buf []byte) (err error) {
		if (*c)[indx] == nil {
			(*c)[indx] = new(Contribution)
		}
		if err = (*c)[indx].UnmarshalSSZ(buf); err != nil {
			return err
		}
		return nil
	})
}

// MarshalSSZTo --
func (c *Contributions) MarshalSSZTo(buf []byte) (dst []byte, err error) {
	dst = buf
	if size := len(*c); size > 13 {
		return nil, ssz.ErrListTooBigFn("ConsensusData.SyncCommitteeContribution", size, 13)
	}

	offset := 4 * len(*c)
	for ii := 0; ii < len(*c); ii++ {
		dst = ssz.WriteOffset(dst, offset)
		offset += (*c)[ii].SizeSSZ()
	}

	for ii := 0; ii < len(*c); ii++ {
		if dst, err = (*c)[ii].MarshalSSZTo(dst); err != nil {
			return
		}
	}
	return dst, nil
}

// MarshalSSZ --
func (c *Contributions) MarshalSSZ() ([]byte, error) {
	return ssz.MarshalSSZ(c)
}

// SizeSSZ returns the size of the serialized object.
func (c Contributions) SizeSSZ() int {
	size := 0
	for _, elem := range c {
		size += 4
		size += elem.SizeSSZ()
	}
	return size
}
